Splitting camFusion_Student.cpp into specific task files.
