+++ CMakeLists.txt
@@ -14,5 +14,14 @@ link_directories(${OpenCV_LIBRARY_DIRS})
 add_definitions(${OpenCV_DEFINITIONS})
 
 # Executable for create matrix exercise
-add_executable (3D_object_tracking src/camFusion_Student.cpp src/FinalProject_Camera.cpp src/lidarData.cpp src/matching2D_Student.cpp src/objectDetection2D.cpp)
+add_executable (
+  3D_object_tracking
+  src/camFusion_Student.cpp
+  src/camFusion_bbmatch.cpp
+  src/camFusion_ttccamera.cpp
+  src/camFusion_cluskpt.cpp
+  src/FinalProject_Camera.cpp
+  src/lidarData.cpp
+  src/matching2D_Student.cpp
+  src/objectDetection2D.cpp)
 target_link_libraries (3D_object_tracking ${OpenCV_LIBRARIES})
