+++ src/camFusion_Student.cpp
@@ -135,250 +135,6 @@ void show3DObjects(std::vector<BoundingBox> &boundingBoxes, cv::Size worldSize, 
 }
 
 
-
-// associate a given bounding box with the keypoints it contains
-void clusterKptMatchesWithROI (BoundingBox &boundingBox, std::vector<cv::KeyPoint> &kptsPrev, std::vector<cv::KeyPoint> &kptsCurr, std::vector<cv::DMatch> &kptMatches)
-{
-    // The objective is to complete
-    //     boundingBox.kptMatches;
-    // defined as
-    //     std::vector<cv::DMatch> kptMatches;
-    // This vector is the subset of all keypoint matches that belong to
-    //     the bounding box in question
-    // Outliers are to be removed
-    
-    std::vector<cv::DMatch> allmatches; // All possible matches, pre-pruning
-    std::vector<float> euclidean; // List of euclidean distances of individual matches
-    
-    float sumLengths = 0.0f;   // Running sum of all euclidean distances
-    float shortest = 1E10;   // Also note shortest and longest
-    float longest = 0.0f;
-
-    bool verbose = true;
-    
-    ///////////////////////////////////////////////////////////////////
-    /// Cycle over all keypoint matches
-    //
-    for (auto it=kptMatches.begin(); it!= kptMatches.end(); it++) {
-	
-	cv::DMatch match = *it;
-	int prevId = match.queryIdx; // IMPORTANT: query == previous !
-	int currId = match.trainIdx; // IMPORTANT: train == current !
-	int imgIdx = match.imgIdx;
-	
-	/////////////////////////////////////////////////////////////
-	/// Keep those that are inside 'boundingBox'
-	//
-	cv::Point2f currPt = kptsCurr[currId].pt;
-	if (boundingBox.roi.contains (currPt)) { 
-	    
-	    allmatches.push_back (*it);  // Inside the bounding box!
-	    
-	    ////////////////////////////////////////////////////////
-	    /// Monitor the spread of match lengths
-	    //
-	    cv::Point2f prevPt = kptsPrev[prevId].pt;
-	    float length = cv::norm (currPt - prevPt);
-	    euclidean.push_back (length);  // Note the euclidean distance 
-	    
-	    sumLengths += length; // Maintain sum, min & max
-	    if (length > longest ) { longest = length; }
-	    if (length < shortest ) { shortest = length; }
-	}
-    }
-    
-    // Sanity check
-    if (allmatches.size() != euclidean.size()) {
-	exit (-1);
-    }
-    
-    if (allmatches.size() == 0) {
-	cout << "Fault in clusterKptMatchesWithROI!\n";
-	exit(-1);
-    }
-
-    float avgEuclid = sumLengths/allmatches.size();
-    
-    if (verbose) {
-	cout << "\nclusterKptMatchesWithROI:\n";
-	cout << "AllMatches=" << kptMatches.size() << " possible=" << allmatches.size() << endl;
-	cout << "Pre: longest=" << longest << " shortest=" << shortest << "  avg=" << avgEuclid << endl;
-    }
-    
-    
-    ///////////////////////////////////////////////////////////////////
-    /// Prune away outliers
-    //
-    float factor = 0.1;    // I set pruning factor to '0.4'
-    float minimumE = factor*avgEuclid;
-    float maximumE = (1.0 + factor)*avgEuclid;
-
-    // testing 
-    minimumE = 0.005*avgEuclid;
-    maximumE = 1.5*avgEuclid;
-
-    shortest = 1E10; // For fun
-    longest = 0.0f;
-    float sum = 0.0;
-    
-    for (ushort n=0; n<allmatches.size(); n++) {
-	float distance = euclidean[n];
-	if (distance >= minimumE && distance <= maximumE) { // Within limits?
-	    boundingBox.kptMatches.push_back (allmatches[n]); // Yes
-
-	    sum += distance;
-	    if (distance > longest ) { longest = distance; }
-	    if (distance < shortest ) { shortest = distance; }
-	}
-    }
-
-    if (verbose) {
-	int survivors = boundingBox.kptMatches.size();
-	cout << "Post: longest=" << longest << " shortest=" << shortest << "  avg="
-	     << sum/survivors << " " << survivors << " survivors\n";
-    }
-    // ...
-    return;
-}
-
-
-// Compute time-to-collision (TTC) based on keypoint correspondences in successive images
-void computeTTCCamera (std::vector<cv::KeyPoint> &kptsPrev, std::vector<cv::KeyPoint> &kptsCurr, 
-                      std::vector<cv::DMatch> kptMatches, double frameRate, double &TTC, cv::Mat *visImg)
-{
-    // The objective is calculate the TTC using change
-    //  in a 'characteristic' dimension between two images.
-    // The characteristic dimension is the ratio of matched keypoints,
-    //  computed for each image
-    // Outliers are to be remved
-    
-    double minLength = 100.0; // Using same value as in lesson #2
-    std::vector<double> charDim;
-
-    double smallest = 1E10;
-    double largest = 0.;
-    double sum = 0.0;
-    
-    //////////////////////////////////////////////////////////////////////////
-    /// Cycle over the keypoints for the object
-    //
-    for (auto itA=kptMatches.begin(); itA!=kptMatches.end();  itA++) {
-	
-	///////////////////////////////////////////////////
-	/// Anchor
-	//
-	cv::KeyPoint kpCurr = kptsCurr.at (itA->trainIdx);
-	cv::KeyPoint kpPrev = kptsPrev.at (itA->queryIdx);
-	
-	for (auto itB=kptMatches.begin()+1; itB!=kptMatches.end();  itB++) {
-	    
-	    //if (itB == itA) { continue; } // Skip the identical match
-	    
-	    ///////////////////////////////////////////////
-	    /// Reference
-	    //
-	    cv::KeyPoint kpCL = kptsCurr.at (itB->trainIdx);
-	    cv::KeyPoint kpPL = kptsPrev.at (itB->queryIdx);
-	    
-	    double lenCurr = cv::norm (kpCurr.pt - kpCL.pt);
-	    double lenPrev = cv::norm (kpPrev.pt - kpPL.pt);
-	    
-	    // Guard against division-by-zero and beyond-precision numbers.
-            if (lenPrev > std::numeric_limits<double>::epsilon() && lenCurr >= minLength) {
-                double distRatio = lenCurr / lenPrev;
-                charDim.push_back (distRatio);
-
-		sum += distRatio;
-		if (distRatio > largest) { largest = distRatio; }
-		if (distRatio < smallest) { smallest = distRatio; }
-	    }
-	    
-	}
-    }
-
-    // Sanity check
-    if (charDim.size() == 0) {
-	TTC = NAN;
-	return;
-    }
-
-    // Statistics
-    std::sort (charDim.begin(),charDim.end());
-    double sumUp = std::accumulate (charDim.begin(),charDim.end(), 0.0);
-    int count = charDim.size();
-    double avgRatio = sumUp / count;
-    double median = charDim[floor(charDim.size()/2)];
-    
-    bool verbose = true;
-    if (verbose) {
-	cout << "\ncomputeTTCCamera:\n";
-	cout << "Pre vec: Small=" << charDim[0] <<  " larg=" << charDim[count-1]
-	     << " avg=" << avgRatio << " @ " << count
-	     << " median=" << median << endl;
-	for (auto it=charDim.begin();it!=charDim.end(); it++) {
-	    //cout << *it << endl;
-	}
-    }
-
-
-    ///////////////////////////////////////////////////////////////////
-    /// Prune away outliers
-    //
-    double referenceDistance;
-    if (false) {
-
-	//**********************************************************
-	// VERY surprising result that 'median' is better
-	// than an 'average'. Indicates (to me) that this method
-	// of estimating TTC is 'touchy'
-
-	float factor = 0.75;    // Acceptance range
-	float minimumR = factor*avgRatio;
-	float maximumR = (1.0 + factor)*avgRatio;
-	
-	count = 0;  // Keep track of pruning effect
-	sumUp = 0.0;
-	smallest = 1E10;
-	largest = 0.0;
-	
-	for (auto it=charDim.begin();it!=charDim.end(); it++) {
-	    double ratio = *it;
-	    if (ratio <= maximumR && ratio >= minimumR) {
-		sumUp += ratio;
-		
-		count++;
-		if (ratio > largest) { largest = ratio; }
-		if (ratio < smallest) { smallest = ratio; }
-	    }
-	}
-	// Adjusted average ration w/outliers removed
-	avgRatio = sumUp / count;
-	referenceDistance = avgRatio;
-
-    } else {
-	// Adjust for an even count
-	if ( (charDim.size() % 2) == 0 ) {
-	    int m = floor(charDim.size())/2;
-	    median = (charDim[m-1]+charDim[m+0])/2;
-	}
-	referenceDistance = median;
-    }
-
-
-    // Estimated time to collision (TTC)
-    double dT = -1.0 / frameRate;
-    TTC = dT / (1.0 - referenceDistance);
-
-    if (verbose) {
-	cout << "Post vec: Small=" << smallest <<  " larg=" << largest
-	     << " avg=" << avgRatio << " @ " << count << " TTC=" << TTC
-	     << " frame rate=" << frameRate << endl;
-    }
-
-    // ...
-}
-
-
 void computeTTCLidar(std::vector<LidarPoint> &lidarPointsPrev,
                      std::vector<LidarPoint> &lidarPointsCurr, double frameRate, double &TTC)
 {
@@ -455,110 +211,6 @@ void computeTTCLidar(std::vector<LidarPoint> &lidarPointsPrev,
     }
     return;
 }
-
-
-void matchBoundingBoxes(std::vector<cv::DMatch> &matches, std::map<int,int> &bbBestMatches, DataFrame &prevFrame, DataFrame &currFrame)
-{
-    // The objective is to return a 'map' relating
-    // the ID's of bounding boxes of two frames
-    // representing the same object
-    // 
-    // We will create a 2-dim matrix of box ID's
-    // first index for previous, second for current.
-    // Entries in the matrix represent number of occurrences
-    // of matched keypoints occupying the indicated box IDs.
-    //
-    // For each 'row' i.e. previous frame boxes, we select the
-    // 'column' i.e. current frame boxes with the most occurrences
-    // or number of matching keypoints
-
-
-    ////////////////////////////////////////////////////////////////////////
-    /// Create and initialize/zero the Matrix
-    //
-    int cPrevBBs = prevFrame.boundingBoxes.size();
-    int cCurrBBs = currFrame.boundingBoxes.size();
-
-    std::vector<std::vector<int> > vPrCr;
-    for (int p=0; p<cPrevBBs;p++) {
-	std::vector<int> vC;
-	for (int c=0; c<cCurrBBs;c++) {
-	    vC.push_back(0);
-	}
-	vPrCr.push_back (vC);
-    }
-
-    /////////////////////////////////////////////////////////
-    /// Call keypoints out to save some typing
-    //
-    std::vector<cv::KeyPoint> prevKpts = prevFrame.keypoints;
-    std::vector<cv::KeyPoint> currKpts = currFrame.keypoints;
-
-
-    //////////////////////////////////////////////////////////////////
-    /// Cycle over all keypoint matches
-    //
-    for (auto itM=matches.begin();itM!=matches.end();itM++) {
-
-	// Process the current match	
-	cv::DMatch match = (*itM);
-	cv::KeyPoint kpPrev = prevKpts[match.queryIdx];
-	cv::KeyPoint kpCurr = currKpts[match.trainIdx];
-
-	//////////////////////////////////////////////////////////////
-	/// Cycle over previous frame's bounding boxes
-	//  to find its keypoints (prev keypoints)
-	//
-	for (auto ibP=prevFrame.boundingBoxes.begin();ibP!=prevFrame.boundingBoxes.end();ibP++) {
-	    if (ibP->roi.contains(kpPrev.pt) ) {
-		int preBoxID = ibP->boxID; // Note the previous box ID
-		
-		//////////////////////////////////////////////////////////////
-		/// Cycle over current frame's bounding boxes
-		//  for a match
-		//
-		for (auto ibC=currFrame.boundingBoxes.begin();ibC!=currFrame.boundingBoxes.end();ibC++) {
-		    if (ibC->roi.contains(kpCurr.pt)) {
-			int curBoxID = ibC->boxID;
-			vPrCr[preBoxID][curBoxID] += 1; // Up the count
-		    }
-		}
-
-	    }
-	}
-    }
-
-    ///////////////////////////////////////////////////////////////
-    /// Cycle over the Matrix to pick out the maximum count
-    /// for each row
-    //
-    for (auto itP=vPrCr.begin();itP!=vPrCr.end();itP++){
-	
-	int prevBoxID = itP-vPrCr.begin(); // Match this box ID
-	
-	int max = 0;  // Initialize the search
-	int currBoxID = 0;
-	std::vector<int> vC = (*itP);
-
-	///////////////////////////////////////////////////////////
-	/// Look for maximum keypoints count in the row of
-	//  matched boxes
-	//
-	for (auto itC=vC.begin();itC!=vC.end();itC++){
-	    int boxID = itC-vC.begin();
-
-	    if ( (*itC) > max ) {
-		max = (*itC);
-		currBoxID = boxID; // Best match currently
-	    }
-	}
-
-	// Best match
-	bbBestMatches.insert (std::pair<int,int>(prevBoxID, currBoxID));
-    }
-    return;
-}
-
 
 
 // Purpose: Return a vector of KeyPoints contained w/i the
